Langchain_App/requirements.txt
langchain
langchain-openai
groq
python-dotenv